Fix long outstanding bug with LADSPA plugin parameter adjustments.

+++ src/tX_seqpar.cc
@@ -953,7 +953,7 @@ const char * tX_seqpar_vttfx :: get_name()
 
 void tX_seqpar_vttfx_float :: create_widget()
 {
-	float tmp=max_value - min_value/1000;
+	double tmp=(max_value - min_value)/1000.0;	
 
 	*fx_value=min_value;
 	//myadj=GTK_ADJUSTMENT(gtk_adjustment_new(*fx_value, min_value+tmp/10, max_value-tmp/10, tmp, tmp, tmp));
